requirements.txt
flask==3.0.3
torch>=2.6.0,<2.10
torchvision>=0.21.0,<0.25
pillow==10.4.0

